Add AWS CDK to dependencies and begin using it
