+++ .gitignore
@@ -1,1 +1,3 @@
 run
+node_modules
+cdk.out
