+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "aws-cdk": "^2.14.0"
+  }
+}
