+++ src/cdk.json
@@ -1,0 +1,3 @@
+{
+  "app": "python3 substrate.py tapestry"
+}
