+++ src/substrate.py
@@ -1,175 +1,60 @@
 #!/usr/bin/env python3
 from argparse import ArgumentParser
 import os
-from pathlib import Path
 import signal
-from urllib.request import urlretrieve
-from urllib.parse import urlparse
+import sys
 
-from docker import from_env
-import paramiko
+from aws_cdk import App
 import yaml
 
-from modules import Tapestry, VCI
-
-
-MODULES = {
-	'tapestry': Tapestry,
-	'vci': VCI
-}
+from classes import SubstrateStack, SubstrateSwarm
 
 
 class Substrate():
-	def __init__(self, tool_name, path=None):
-		self.docker = from_env()
-		self.network = None
+	def __init__(self, _args):
+		config = self.parse_yaml(_args.path)
 
-		self.config_name = 'substrate.config.yaml'
-		self.config = self.parse_yaml(path)
+		if config.get('aws', None):
+			os.chdir(os.path.dirname(os.path.realpath(__file__)))
 
-		data_path = self.get_data()
+			app = App()
+			SubstrateStack(app, 'substrate-stack', config=config)
+			app.synth()
 
-		self.tool_name = tool_name
-		if self.tool_name not in MODULES:
-			raise Exception(f'No tool named {self.tool_name}')
-		self.tool = MODULES[self.tool_name](self.docker, self.config, data_path)
+		elif config.get('cluster', None):
+			substrate = SubstrateSwarm(_args.tool, config)
 
-		ssh_dir = os.path.join(Path.home(), '.ssh')
-		ssh_dirfiles = os.listdir(ssh_dir)
-		ssh_keys = [keyfile for keyfile in ssh_dirfiles if keyfile.startswith('id_')]
-		ssh_pkeys = [keyfile for keyfile in ssh_keys if not keyfile.endswith('.pub')]
-		self.key_paths = [os.path.join(ssh_dir, keyfile) for keyfile in ssh_pkeys]
-
-	def create_swarm(self):
-		advertise_addr = self.config['cluster'].get('advertise_addr', '0.0.0.0')
-		self.docker.swarm.init(advertise_addr=advertise_addr)
-
-		self.network = self.docker.networks.create(
-			f'substrate-{self.tool_name}-net',
-			driver='overlay'
-		)
-
-		manager_token = self.docker.swarm.attrs['JoinTokens']['Manager']
-		worker_token = self.docker.swarm.attrs['JoinTokens']['Worker']
-
-		nodes = []
-		for node in self.config['cluster'].get('managers', []):
-			nodes.append(('manager', node))
-		for node in self.config['cluster'].get('workers', []):
-			nodes.append(('worker', node))
-
-		for node_type, node in nodes:
-			username, location = node.split('@')
-			if node_type == 'manager':
-				token = manager_token
-			elif node_type == 'worker':
-				token = worker_token
-
-			self.log(f'Adding remote {node} to swarm as {node_type}…')
-
-			with paramiko.client.SSHClient() as ssh_client:
-				ssh_client.load_system_host_keys()
-				ssh_client.connect(
-					location,
-					username=username,
-					key_filename=self.key_paths
-				)
-
-				_, _, stderr = ssh_client.exec_command(
-					f'docker swarm join --token {token} {advertise_addr}'
-				)
-				stderr = stderr.read()
-				if stderr:
-					self.log(f'✕\nremote error ({location}): {stderr}\n')
-				else:
-					self.log('✓\n')
-
-		# TODO: support AWS
-
-	def destroy_swarm(self):
-		nodes = []
-		for node in self.config['cluster'].get('managers', []):
-			nodes.append(node)
-		for node in self.config['cluster'].get('workers', []):
-			nodes.append(node)
-
-		for node in nodes:
-			username, location = node.split('@')
-
-			self.log(f'Removing remote {node} from swarm…')
-
-			with paramiko.client.SSHClient() as ssh_client:
-				ssh_client = paramiko.client.SSHClient()
-				ssh_client.load_system_host_keys()
-				ssh_client.connect(
-					location,
-					username=username,
-					key_filename=self.key_paths
-				)
-
-				_, _, stderr = ssh_client.exec_command('docker swarm leave --force')
-				stderr = stderr.read()
-				if stderr:
-					self.log(f'✕\nremote error ({location}): {stderr}\n')
-				else:
-					self.log('✓\n')
-
-		self.docker.swarm.leave(force=True)
-
-	def get_data(self):
-		source_path = self.config['data']['source']
-
-		# Download the dataset if necessary to the target location
-		if urlparse(source_path).scheme != '':
-			target_path = os.path.abspath(self.config['data']['target'])
-
-			os.makedirs(os.path.dirname(target_path), exist_ok=True)
-			urlretrieve(source_path, target_path)
-
-			data_path = target_path
-		else:
-			data_path = os.path.abspath(source_path)
-
-		if os.path.isfile(data_path):
-			data_path = os.path.dirname(data_path)
-
-		return data_path
-
-	# TODO: use more sophisticated logging setup
-	def log(self, message):
-		print(message, end='')
+			substrate.start()
+			signal.sigwait([signal.SIGINT])
+			substrate.stop()
 
 	def parse_yaml(self, path):
+		config_name = 'substrate.config.yaml'
+
 		if path is None:
 			path = os.getcwd()
 
 			files = os.listdir(path)
-			while self.config_name not in files:
+			while config_name not in files:
 				path = os.path.abspath(os.path.join(path, '..'))
 				files = os.listdir(path)
 
 				if path == '/':
 					break
 
-			if self.config_name not in files:
-				raise FileNotFoundError(self.config_name)
+			if config_name not in files:
+				raise FileNotFoundError(config_name)
 
-			path = os.path.join(path, self.config_name)
+			path = os.path.join(path, config_name)
 
 		with open(path, 'r', encoding='utf8') as stream:
-			return yaml.load(stream, Loader=yaml.Loader)
+			_config = yaml.load(stream, Loader=yaml.Loader)
 
-	def start(self):
-		self.create_swarm()
-		self.log('Initialized Swarm.\n')
-		self.tool.start()
-		self.log(f'Started {self.tool_name}. Press Ctrl+C to exit.\n')
+		# TODO: add as much error checking as I have time for
+		if _config.get('cluster', None) and _config.get('aws', None):
+			sys.exit('The "cluster" and "aws" options cannot be used simultaneously.')
 
-	def stop(self):
-		self.tool.stop()
-		self.log(f'Stopped {self.tool_name}.\n')
-		self.destroy_swarm()
-		self.log('Destroyed Swarm.\n')
+		return _config
 
 
 if __name__ == '__main__':
@@ -183,8 +68,4 @@ if __name__ == '__main__':
 	parser.add_argument('-c', '--config', dest='path')
 	args = parser.parse_args()
 
-	substrate = Substrate(args.tool, path=args.path)
-
-	substrate.start()
-	signal.sigwait([signal.SIGINT])
-	substrate.stop()
+	Substrate(args)
